refactor(artwork): address CR — shared normalizer to artwork_format, sync→data edge documented
KHPilot findings on #170 plus converging independent-CR nits:
- normalized_artwork_form_name moves from artwork_local to artwork_format
  (the shared module) so the two backends stay peer-independent
  (independent CR + bot both flagged the peer-backend coupling).
- STYLE documents the third deliberate layer exception: sync/images_sync
  consumes data/images' AKDP index contract (parse_index/ImagesIndex/
  SCHEMA_VERSION) — the index is the sync-output/data-input contract and
  relocating it would invert into a worse data → sync edge.
- Dead TS imports from the atomicWriteJson dedup removed (gamedataPair:
  randomUUID/basename/writeFile; releaseActivation: readdir — the PY twin
  had dropped its uuid4).
- prune_generations' parameter renamed tree_root to stop shadowing the
  releases_dir() module function (both sides).
